convert wakeup fifo reads to a dedicated fd
This also silences errors in shells that fail to 'read <fifo' while it
is being written to by another process (dash 0.5.8.2, mksh 52).

The errors would be printed even when redirecting 'read' as '>/dev/null 2>&1',
perhaps because the shell cannot open the fifo, and opening
belongs to the shell itself, not the 'read' builtin.

Possibility for the errors being printed out was introduced with the
sync daemon, but has only been observed after converting the parallel
each to use it.